move setup server file out of exp
