+++ packages/cli/src/commands/setup/server-file/templates/server.ts.template
